src/services/optimization/TransportationOptimizer.ts
import { RouteSegment } from '@/types/route/types';
import { GeoPoint } from '@/types/geo';
import { ActivityType } from '@/types/activity';
import { UserPreferences } from '@/types/ai/learning';
import logger from '@/utils/logger';

type TransportMode = 'walk' | 'bike' | 'car' | 'bus' | 'train' | 'ski';

interface TransitConnection {
  type: TransportMode;
  startPoint: GeoPoint;
  endPoint: GeoPoint;
  schedule: {
    frequency: number;
    firstDeparture: string;
    lastDeparture: string;
  };
  restrictions?: {
    capacity: number;
    accessibility: string[];
    equipment: string[];
  };
}

export class TransportationOptimizer {
  async optimizeMultiModal(
    startPoint: GeoPoint,
    endPoint: GeoPoint,
    preferences: UserPreferences
  ): Promise<RouteSegment[]> {
    try {
      // Get available transport modes
      const availableModes = await this.getAvailableTransportModes(startPoint, endPoint);
      
      // Get transit connections
      const connections = await this.getTransitConnections(startPoint, endPoint, availableModes);

      // Compute optimal route
      return this.computeOptimalRoute(startPoint, endPoint, availableModes, connections, preferences);
    } catch (error) {
      logger.error('Failed to optimize multi-modal route:', error);
      throw error;
    }
  }

  private async getAvailableTransportModes(
    startPoint: GeoPoint,
    endPoint: GeoPoint
  ): Promise<TransportMode[]> {
    // Query available transport options in the area
    return [
      'walk',
      'bike',
      'car',
      'bus',
      'train',
      'ski'
    ];
  }

  private async getTransitConnections(
    startPoint: GeoPoint,
    endPoint: GeoPoint,
    modes: TransportMode[]
  ): Promise<TransitConnection[]> {
    // Query transit connections between points
    return [];
  }

  private async computeOptimalRoute(
    startPoint: GeoPoint,
    endPoint: GeoPoint,
    modes: TransportMode[],
    connections: TransitConnection[],
    preferences: UserPreferences
  ): Promise<RouteSegment[]> {
    const route: RouteSegment[] = [];
    let currentPoint = startPoint;

    while (!this.isAtDestination(currentPoint, endPoint)) {
      const nextSegment = await this.findOptimalSegment(
        currentPoint,
        endPoint,
        modes,
        connections,
        preferences
      );

      if (!nextSegment) {
        throw new Error('Unable to find valid route segment');
      }

      route.push(nextSegment);
      currentPoint = nextSegment.endPoint;
    }

    return this.optimizeRouteTransitions(route, preferences);
  }

  private async findOptimalSegment(
    startPoint: GeoPoint,
    endPoint: GeoPoint,
    modes: TransportMode[],
    connections: TransitConnection[],
    preferences: UserPreferences
  ): Promise<RouteSegment | null> {
    // Implementation to find optimal segment
    return null;
  }

  private isAtDestination(current: GeoPoint, destination: GeoPoint): boolean {
    const threshold = 0.1; // 100 meters threshold
    return (
      Math.abs(current.lat - destination.lat) < threshold &&
      Math.abs(current.lng - destination.lng) < threshold
    );
  }

  private async optimizeRouteTransitions(
    route: RouteSegment[],
    preferences: UserPreferences
  ): Promise<RouteSegment[]> {
    // Optimize transitions between segments
    return route;
  }
} 